upload_misc_fields: split InputParser logic up into three parts, hopefully making each part easier to digest
Unfortunately, because the scala Parser framework wants all your code to be part of a
particular trait, I have had to split the code by making each part a layer in a stack of
traits.
